Change registry value for config directory to autostart_config_dir
This allows to isolate autostarted configs from those accessed
through the GUI into separate directories.

This also allows setting the service start type to default to Auto
(not manual) without unexpectedly starting profiles in config_dir.

If the registry value does not exist or is empty it is ignored. Otherwise
it must point to a directory that exists. The installer should set up
a default value and default directory.

See
https://github.com/OpenVPN/openvpn-build/pull/141#issuecomment-442536134
and subsequent discussion.

Signed-off-by: Selva Nair <selva.nair@gmail.com>